CI(azure): Fix usage of BUILD_NUMBER_TOKEN
Turns out environment variables cannot be exported globally:

"/.ci/azure-pipelines/main.yml (Line: 7, Col: 1): Unexpected value 'env'"

It has to be done for each step that makes use of the variable.

Furthermore if a variable is not defined on Azure, the expansion
$(MY_VARIABLE) does not resolve to an empty String (as is the case e.g.
in Bash) but this expression is then taken literally without any
expansion at all. Therefore we have to make sure that the
BUILD_NUMBER_TOKEN variable is always defined. Thus we define it in
main-pr.yml as for PRs we don't define it on Azure.

In order for these variables to not interfere with each other, the
variable name inside the scripts was prefixed with MUMBLE_