REFAC(client): Move CompletablePage class into dedicated file set
The class was defined in AudioWizard.cpp/.h, however AudioWizard is not the only user.

In fact, Cert.ui had to include AudioWizard.h in order to make use of CompletablePage.

This commit moves the class into a dedicated file set and changes the include header in AudioWizard.ui and Cert.ui.