CI(azure): Use version & commit hash as identifier for build number
Our script on the server now bumps the build number for every new version + commit combination.
Each combination is associated to a unique build number and vice versa.

The previous system didn't cause collisions, but it was very fragile.
For example, the build number got de-synced between OSes a few days ago.
It probably happened due to at least a build being restarted after the build number bump.

With the new system there should be no problems anymore, as the build number is permanent for a given identifier.