FIX(client): Restore default echo cancellation option
In #4694 the echo cancellation settings were refactored. During this,
the default value was changed from being Speex's mixed channel mode to
being off.

This commit restores the previous default except for macOS where Speex
echo cancellation does not work. There Apple's echo cancellation system
is set as the default. Although that only works with a certain
combination of input and output devices, it doesn't make the situation
worse for other combinations, so we might as well enable it by default.