FIX(client): Default echo cancel setting not loaded correctly
613a5d5fb0ed4078e3a127b784eada14a809fa50 introuced a compatibility layer
for porting over the old echo cancel settings to the new one. However by
doing so it prevented the default setting (Speex mixed) from being
loaded properly as in this case it is not saved explicitly. This would
lead to the compat block being executed that would disable echo
cancelling if there are no old settings.

This commit fixes this by making sure the compat block is only run when
the old setting is actually present.

Fixes #4761