FEAT(client): Changed shorcut for channel filtering
The shortcut for toggling channel filters was Ctrl+F. While this makes
sense semantically, this specific shortcut is associated with starting a
search in 99% of all applications out there. Thus using it for something
else does not make for a very smooth user experience.

The new shortcut for filtering is thus set to be Alt+F instead.