BUILD(cmake): Fix pkg search on case-insensitive filesystems
If a package is searched for with multiple casings (as is currently the
case for libsndfile), it will be very hard to use the resulting cmake variables
(e.g. <name>_FOUND, <name>_LIBRARIES) as one does not know which casing was
eventually found.

This issue actually broke the build on macOS when installing libsndfile
via homebrew.

This change sets the all-lowercase version of the package
name variables (if multiple cases of the spelling are contained in the
package's aliases) so that this can then be used further down to reference
the result, without needing to iterate through all the options again to
determine which case was found.

Fixes #4790