FEAT(server): Remove broadcast of new welcometext when changing via Ice.
When changing the welcome text via Ice, Murmur would broadcast the new text to all connected clients. It is a matter of opinion if this is proper behavior or not, but it seems too proactive of an action. Let the clients see the new welcometext when they next connect, or let the Murmur administrator explicitly perfor a sendMessage() to all connected clients after changing the welcometext.

Closes: #4788

Co-Authored-By: Robert Adam <dev@robert-adam.de> <krzmbrzl@gmail.com>
Co-Authored-By: Robert Adam <dev@robert-adam.de> <krzmbrzl@users.noreply.github.com>