MAINT: Update updatetranslation.sh script to update all ts files from sources
Our previous workflow was using `scripts/updatetranslations.sh` to update `src/mumble/mumble_en.ts` only.

We then depended on Transifex to take over text source changes from this file to translation files.

This lead to committed changes being a mix of both translation source changes as well as translations.
This made the diffs unnecessarily big, and hard to review as two concerns are intermixed.

Weblate supports this style of updating as well. It calls this style 'monolingual', although that concept is intermixed with, or designed for files that do not contain translation source text (which ours do).
One file is used as a source, and another contains the translation strings.

Seemingly Qt calls this approach text-by-ID, and the translation strings are looked up by the ID, which in our case are the translation source strings.
This seems like a misuse of the concept.

The monolingual vs bilingual translation files issue/discrepancy is not handled in this change.

---

From now on we update all ts files so an update to the translation files with this script leads to an update in all translation files
rather than only the 'mumble_en.ts' reference file.

This approach will make translation change commits and PRs afterwards isolated to translation changes, and thus a lot smaller and easier to review.
There will no longer be an invisible, implicit difference between the committed ts file and the current program state = the translation source.