MAINT: Drop ts file plural fixup in updatetranslations.sh
Text with plural forms (a %n parameter and singular(s) qualifier) lead to special plural handling.

Different locales can have varying words depending on the number; mostly partially differently between 0, 1 and more.

English texts can have simple 'word(s)' text which will be transformed to '0 word', '1 word', '2 words', '3 words' and so on respectively.

Other languages may not have this mechanism, and the ts files are prepared with `<numerus>` elements within the `<translation>` elements.
These seem to be for specifying the varying plural forms manually.

This seems to be an automatism by lupdate.
When updating our current ts files it warns about multiple instances of the number of `numerus` elements not matching the number of plural forms in the locale the file is for.

Whatever the original hack/fix was for (#1195, the linked Transifex topic no longer works),
it seems very problematic.

Previously we only updated mumble_en.ts, and then depended on Transifex to update the contents of the other ts files when translations are changed.
This led to source and translation changes being intermixed in the change commit.

This change drops the 'numerus' element duplication.
This is in preparation for in the future updating translation sources on all ts files, where this fixup could be technically incorrect and problematic as a result.
As well as to drop this hack which seems like an awful, misleading and technically incorrect workaround for Transifex bugs or side effects in it’s transformation of source ts file to translation ts files.

The 'numerus' tag is not documented in Qt documentation with any information.

https://doc.qt.io/qt-5/i18n-source-translation.html#handling-plurals