MAINT: Added script for updating license headers
The script iterates through (almost) all files within the Mumble
repository and updates the copyright notice in them. If source files
don't contain such a notice yet, it can optionally add one in.

The script asks git when the file was first added to the git index and
uses that year as the starting year for the copyright notice. The end
year is always set to be the current year at the time the script is run.