CI(windows): Increase timeout
The timeout has been increased for the Windows CI builds. The new
timeouts are:
- x64: 90 minutes
- x86: 120 minutes

x86 has the higher timeout because it always takes quite a bit longer
than the x64 build.

The timeout was increased as we frequently see the CI job being
cancelled because it did not finish within the default 60 minutes
timeout.