MAINT: Replace updatetranslations.sh with updatetranslations.py script
The bash script is not platform independent. The python script is.

I invested way too much time trying to run the sh script on WSL (Windows Subsystem for Linux), installing Qt packages, but in the end still failing.

I also implemented a bat script calling lupdate with the same parameters as the sh script, but having one platform independent script is preferable to having two separate scripts for the same activity.

The python script will call lupdate to update the ts files like the sh script did,
and commit the changes if any were made.

Furthermore it checks that the ts files have no local changes before running,
and after updating reset the git index, add commit the files.
The reset operation is added so only the ts files are being committed.

Contrary to the previous script this one disables all heuristics so those can be applied as separate changes, so they will be visible.