FEAT(client): Improve translation directory search
Mumble allows using translation files (*.qm) on disk to overwrite the
bundled translations. This is very useful for translators wanting to try
out their translations.

Unfortunately though in order for this to work translations had to be
placed next to the Mumble executable which (due to the usual install
path of Mumble) requires admin/root privileges.

Therefore this commit adds a few extra locations in which such
translation files are searched in.

Furthermore it adds the --translation-dir command line argument that
allows to specify a custom location that is prepended to the list of
checked directories.

Fixes #4796