BUILD(cmake): Use big file linker option
Using a unity build tends to create large source files. With that it is
easily possible to hit the object file's size limit resulting in the
"C1128: number of sections exceeded object file format limit" (MSVC) or
"File too big" (GCC) error.

In order to avoid this, this commit sets the necessary flag to extend
that limit.