BUILD(cmake): Exclude files from unity build
The *_plugin_import files will cause collisions in a unity build, so we
have to make sure that they are built separately as usual.

Furthermore the ObjecticeC files don't play well in unity builds (it
seems like they are then treated as cpp files). Thus they are excluded
from that as well.

WinGUID.cpp must also not be included as that breaks the DEFINE_GUID
macro wizardry needed to avoid redefinition errors.