TRANSLATION: Revert "REFAC(client): Use QObject to access static tr()"
This reverts commit 591741c9a6588107431f68e81f34ad64af65d0fd
which was part of PR #4777.

That change switched `tr` calls from the specific class to being referenced on QObject.
That broke all of the translations we have for these strings.

Qt/tr assigns a translation context, which is the Q_OBJECT context by default. By not calling tr on the specific QObject but the base function, the context is dropped. The strings are now no longer in the ConnectDialog but the generic QObject.

Obviously that is undesired.

Reverting that change is the first step to manually restoring the translations.
They were dropped automatically by updating translations - and the previous strings no longer existing.