DOCS(client, windows): Macro wizardry around DEFINE_GUID
There is some preprocessor magic going on for the DEFINE_GUID macro that
causes it to only declare an extern variable (instead of defining and
initializing and actual variable) unless the initguid.h header was used.
The latter is done in WinGUIDs.cpp in order to really initialize the
variables somewhere whereas everywhere else (namely WASAPI.cpp) only
references these extern variables.