FEAT(client): Search dialog
This commit introduces a fully featured search dialog.

Note that selections in the search dialog are synchronized with the main
UI and thus "whisper to selection" works by selecting something in the
search dialog as well.

There are also default actions that are executed when a search result is
activated (double-clicked or pressing enter on it). What that action is
exactly, can be configured in the settings.

Furthermore the context menu works as expected when invoked on entries
in the search result list.

In order to disambiguate the results, the full channel hierarchy to the
search result's parent channel is shown below each result.

The search dialog can be toggled via Ctrl+F when Mumble has focus or by
using the new entry in the toolbar. Additionally there is a new global
shortcut that can be configured for this purpose. Note however that on
Windows the search dialog won't obtain focus when toggled in a situation
when Mumble does not have focus already.