BUILD(cmake): Fix unity build issue
On Unix there are the X11 headers that define incredibly many macros of
which basically none are prefixed in any way. Thus these can easily lead
to name clashes in other files if unity builds are active.

Thus the files known to make problems, are excluded from unity builds in
order to avoid that.