FIX(client, g15): Prevent text on G15 from scaling
The support for rendering text on the Logitech G15's screen uses regular
QFonts for drawing on it. As the font size was specified in points
though, Qt would apply scaling if either a high DPI monitor was in use
or if the user specified a system-wide scaling factor e.g. via Windows
Control Panel.

As the G15 screen is always the same though, we don't want it to scale
and thus we have to set the font's size in pixels instead of in points.
See: https://wiki.qt.io/Technical_FAQ#How_can_I_prevent_the_font_from_scaling_up_when_the_DPI_is_increased.3F

Fixes #4861

Co-Authored-By: Davide Beatrici <git@davidebeatrici.dev>