FIX(client, ui): Ordering of Users inconsistent
If two persons are reading out the users in a channel in the order they
are rendered, chances are that they will read the names in a different
order. This is because the ordering of users takes the currently used
locale into account that can differ from machine to machine.

This is an inconsistency that one probably does not expect.

Therefore this commit changes back to using a lexicographical comparison
that should be the same on all machines, regardless of the locale.

Fixes #4873