FEAT(client): Add User and Listener volume slider to context menus
As requested in #4748 this commit implements a volume control
slider for User local volume and Listener local volume in the
relevant context menus for easier reach and quicker adjustments.

The VolumeSliderWidgetAction was implemented as a base class
for UserLocalVolumeSlider and ListenerLocalVolumeSlider which were
then added as QActions in the MainWindow to the context menus qmUser and
qmListener respectively.

The existing LocalVolumeDialogs for users and listeners have been removed.

Additionally, the new generic MenuLabel class was created to be able to
add non-interactive items into QMenus.

Co-authored-by: Hartmnt <hartmunt@protonmail.com>