FIX(server): Fix undefined behvior on Linux hosts
The fields of the rlimit struct were assumed to be of type long int when
printing via the qWarning function. This assumption is correct for glibc
(used by e.g. Ubuntu) but it is not true for musl (used by e.g. Alpine
Linux).

In that case the wrong format specifiers are being used for printing
these values, which causes undefined behavior.

The fix is to always copy the values into variables of type long long
unsigned int and then printing these. The code also adds an assertion
making sure that within this copy operation the values won't overflow
the variable's size.

Fixes #4897