FIX(client): Memory leak in ALSA implementation
The functions snd_card_get_name and snd_card_get_longname allocate a
String for the caller that needs to be freed once it is no longer needed.
This was not done in the current implementation and thus this was causing
a memory leak.

This commit fixes this by explicitly `free`ing the resources after usage.

Fixes #4910

Co-Authored-By: Zero0one1 <summerlinasity@gmail.com>
Co-Authored-By: Krzmbrzl <krzmbrzl@gmail.com>