FIX(client): Check the returned error code of ALSA
The functions snd_card_get_name and snd_card_get_longname allocate a
String and return 0 if success otherwise the negative error code.
There is a need to check the return value before using the allocated String.

This commit fixes this by checking the return value and exiting if the
return value is not zero.

Ref: #4910

Co-Authored-By: Zero0one1 <summerlinasity@gmail.com>
Co-Authored-By: Krzmbrzl <krzmbrzl@gmail.com>