FIX(client): CrashReporter considering 2xx codes as errors
Apparently nobody has told Qt that 2xx HTTP status codes are considered
success. Therefore checking QNetworkReply::error() in order to check
whether the transfer was successful, does not work.

Instead this commit makes sure that the raw HTTP status code is checked
instead and if it is in range [200,299], the sending of the report is
considered a success.

Furthermore this commit reworks the Strings displayed to the user
(message box) as these were excessively long. These formulations have
been replaced with (hopefully) more concise versions.

Fixes #4928