FIX(client): Auto-connect not working for --multiple
"--config" is parsed before the main argument parsing loop. Then these
arguments are captured a second time in the else clause and shoved
into the "url" variable.

Having a valid url (the path to a config is a valid file URL) causes the
app to post an OpenURLEvent. This is pretty harmless, it just causes the
settings to load again. But the bigger problem is that
Global::get().mw->on_qaServerConnect_triggered(true) is not called. So,
even if you have autoconnect enabled, launching mumble with "--multiple
--config PATH" won't actually autoconnect.

This commit fixes this by skipping over the "--config" arguments in the
main parsing loop.