FEAT(client): Expose flag for default cert path
At startup, the settings for a Mumble instance are loaded (either
using the default location or the "--config" file that was supplied).
Then Mumble checks to see if a valid certificate was loaded. If it
wasn't, it looks for a certificate in a hardcoded path
(QStandardPaths::DocumentsLocation). If there is a certificate there,
it uses it. If not, it creates a new cert and saves it to that
hardcoded path.

This all means that when you run a second instance for the first time,
there will be no certificate in the supplied "--config", so it will use
the certificate in the hardcoded path, which is almost certainly the
certificate that the other mumble instance is using. Connecting to a
server with this certificate will authenticate you as the user who has
that certificate registered. This is probably not what you want.

This commit fixes this issue by adding a CLI option that allows the user
to specify the default certificate directory.