CHANGE(client): Default min volume to 25%
The minimum volume setting determines how much a user's audio will fade
out with increasing in-game distance when positional audio is enabled.
Having this at 80% yields a basically non-noticeable effect leading to a
couple of reports of volume fading not working at all.

Thus this commit changes the default to 25% which should be an audible
difference. It is not set to 0% in order to not run into situations in
which users don't hear each other at all and then thinking that
positional audio is completely broken.

For anyone who wants a realistic audio experience though, this should be
set to 0%.