FEAT(server): Force users to name-match their certificates
It might be desirable to force a user to name him/herself according to the certificate provided to the server. This behaviour is complementary to a client who only trusts a server whichs certificates common name machtes it's hostname.

To achive this, a new ini-option is introduced:
forceUsernameCertSubjectEquality (default: false)
If this is set to true, a user can only enter the server, if her username equals the common-name-subject-field of the provided certificate. If client certificates are not required on a server (set using "certrequired"), this property is ignored.

The check, wheter the name machtes the certificate is done right after an authenticate message is received. For failed checks, a new reject type is introduced: UsernameCertMissmatch.

Implements #4940
Closes #4940