FEAT(server): Add an option to authenticate users purely based on username
If a server can be sure that a user is realy the user which his name says he is, a server could authenticate that user without checking a certificate or password. This is only useful in some cases, mainly these where a server only trusts certificates from certain CAs.

To implement this, a new ini-option is added:
blindTrust (default: false)
This option is very dangerous! If blind trust is set to true, the server trusts the user to be who his username says he is. The only practial case this is useful is if "forceUsernameCertSubjectEquality" is also true, and the MCTS is configured correctly and strongly. However, we won't prevent you from enabeling blindTrust, even if "forceUsernameCertSubjectEquality" is false or your MCTS is not set up correctly, since open source software means freedom, and you are free to do as you please.

Technically this is achived by assuming the user entered the correct password for that user account.

Implements #4961
Closes #4961